+++ grblas/_ss/utils.pyx
