+++ grblas/_ss/vector.py
